refactor(Storybook): Split Table Page stories into individual files
